requirements.txt
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.2
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.4
cycler==0.11.0
decorator==5.1.1
executing==0.8.3
Flask==2.1.2
Flask-Login==0.6.2
Flask-SQLAlchemy==2.5.1
fonttools==4.33.3
google-generativeai==0.3.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==6.0.0
ipython==8.3.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
kiwisolver==1.4.2
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
numpy==1.22.3
packaging==21.3
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.0
plotly==5.8.0
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.8
python-dateutil==2.8.2
python-http-client==3.3.7
pytz==2022.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
sendgrid==6.9.7
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.36
stack-data==0.2.0
starkbank-ecdsa==2.0.3
tenacity==8.0.1
threadpoolctl==3.1.0
traitlets==5.1.1
validate-email==1.3
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.12.0
